# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "一个基于 Gin + GORM + Redis + RabbitMQ 的在线书店后端服务，提供用户认证、书籍管理、购物车、订单及秒杀等功能。",
        "title": "Project1_Shop API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Me",
            "email": "frottenice@outlook.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:9090",
    "basePath": "/",
    "paths": {
...
        "/admin/seckill/create": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "管理员为指定书籍创建秒杀活动，设置秒杀价格、库存和时间（需要管理员权限）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "秒杀管理（管理员）"
                ],
                "summary": "创建秒杀活动（管理员）",
                "parameters": [
                    {
                        "description": "秒杀活动参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.SeckillParam"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "参数错误 / 无权限 / 服务器繁忙",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseData"
                        }
                    }
                }
            }
        },
        "/admin/seckill/down/{id}": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "管理员将指定秒杀活动下架（状态置为 0）（需要管理员权限）",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "秒杀管理（管理员）"
                ],
                "summary": "下架秒杀活动（管理员）",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "秒杀活动 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "参数错误 / 无权限 / 服务器繁忙",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseData"
                        }
                    }
                }
            }
        },
        "/admin/seckill/list": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "管理员获取所有进行中的秒杀活动列表（需要管理员权限）",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "秒杀管理（管理员）"
                ],
                "summary": "获取秒杀活动列表（管理员）",
                "responses": {
                    "200": {
                        "description": "无权限 / 服务器繁忙",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseData"
                        }
                    }
                }
            }
        },
...
        "/api/login": {
            "post": {
                "description": "用户登录，成功后通过 Cookie 下发 access_token 和 refresh_token，同时在响应体中返回 access_token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户"
                ],
                "summary": "用户登录",
                "parameters": [
                    {
                        "description": "登录参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.ParamLogin"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "参数错误 / 用户名或密码错误 / 服务器繁忙",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseData"
                        }
                    }
                }
            }
        },
...
        "/api/register": {
            "post": {
                "description": "注册新用户，需要提供用户名、密码、确认密码、邮箱和性别",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户"
                ],
                "summary": "用户注册",
                "parameters": [
                    {
                        "description": "注册参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.ParamSignUp"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "参数错误 / 用户已存在 / 服务器繁忙",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseData"
                        }
                    }
                }
            }
        },
        "/api/seckill/do": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "登录用户参与指定秒杀活动抢购，成功后异步生成订单（需要登录）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "秒杀"
                ],
                "summary": "参与秒杀",
                "parameters": [
                    {
                        "description": "秒杀参数（product_id 必填）",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.SeckillRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "未登录 / 参数错误 / 活动已结束 / 库存不足 / 重复参与",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseData"
                        }
                    }
                }
            }
        },
...
        "/api/seckill/{id}": {
            "get": {
                "description": "根据秒杀活动 ID 获取详细信息（包含书籍信息、价格、库存、时间等）",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "秒杀"
                ],
                "summary": "获取秒杀活动详情",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "秒杀活动 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "参数错误 / 活动不存在",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseData"
                        }
                    }
                }
            }
        },
...
        "/api/userComments": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "获取当前登录用户发表的所有评论（需要登录）",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户"
                ],
                "summary": "获取当前用户的评论列表",
                "responses": {
                    "200": {
                        "description": "未登录 / 服务器繁忙",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseData"
                        }
                    }
                }
            }
        },
        "/api/userInfo": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "获取当前登录用户的基本信息（需要登录）",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户"
                ],
                "summary": "获取当前登录用户信息",
                "responses": {
                    "200": {
                        "description": "未登录 / 服务器繁忙",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseData"
                        }
                    }
                }
            }
        },
...
        "/api/userRatings": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "获取当前登录用户对书籍的所有评分记录（需要登录）",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户"
                ],
                "summary": "获取当前用户的评分记录",
                "responses": {
                    "200": {
                        "description": "未登录 / 服务器繁忙",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseData"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.ParamLogin": {
            "type": "object",
            "required": [
                "password",
                "username"
            ],
            "properties": {
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "models.ParamSignUp": {
            "type": "object",
            "required": [
                "email",
                "password",
                "re_Password",
                "username"
            ],
            "properties": {
                "email": {
                    "description": "email表示该字段必须是一个合法的邮箱格式",
                    "type": "string"
                },
                "gender": {
                    "type": "integer",
                    "enum": [
                        0,
                        1
                    ]
                },
                "password": {
                    "description": "binding:\"required\"表示该字段为必填项",
                    "type": "string"
                },
                "re_Password": {
                    "description": "eqfield=Password表示该字段必须和Password字段相等",
                    "type": "string"
                },
                "username": {
                    "description": "json表示前端传过来的字段名，binding表示校验规则",
                    "type": "string"
                }
            }
        },
        "models.ResponseData": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "msg": {}
            }
        },
        "models.SeckillParam": {
            "type": "object",
            "required": [
                "book_id",
                "end_time",
                "seckill_price",
                "start_time",
                "stock"
            ],
            "properties": {
                "book_id": {
                    "type": "integer"
                },
                "end_time": {
                    "type": "string"
                },
                "seckill_price": {
                    "type": "integer"
                },
                "start_time": {
                    "type": "string"
                },
                "stock": {
                    "type": "integer",
                    "minimum": 1
                }
            }
        },
...
        "models.SeckillRequest": {
            "type": "object",
            "required": [
                "product_id"
            ],
            "properties": {
                "product_id": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "格式：Bearer {token}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}